Forbid user requests of VBR user

+++ src/main/java/com/tonkar/volleyballreferee/controller/UserGameController.java
@@ -1,6 +1,7 @@
 package com.tonkar.volleyballreferee.controller;
 
 import com.tonkar.volleyballreferee.model.GameDescription;
+import com.tonkar.volleyballreferee.model.UserId;
 import com.tonkar.volleyballreferee.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,18 +25,24 @@ public class UserGameController {
 
     @RequestMapping(value = "", params = { "userId" }, method = RequestMethod.GET)
     public ResponseEntity<List<GameDescription>> listUserGames(@RequestParam("userId") String userId) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         List<GameDescription> games = userService.listUserGames(userId);
         return new ResponseEntity<>(games, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/available", params = { "userId" }, method = RequestMethod.GET)
     public ResponseEntity<List<GameDescription>> listAvailableUserGames(@RequestParam("userId") String userId) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         List<GameDescription> games = userService.listAvailableUserGames(userId);
         return new ResponseEntity<>(games, HttpStatus.OK);
     }
 
     @RequestMapping(value = "", params = { "userId", "kind", "league" }, method = RequestMethod.GET)
     public ResponseEntity<List<GameDescription>> listUserGamesInLeague(@RequestParam("userId") String userId, @RequestParam("kind") String kind, @RequestParam("league") String league) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         league = ControllerUtils.decodeUrlParameter(league);
         List<GameDescription> games = userService.listUserGamesInLeague(userId, kind, league);
         return new ResponseEntity<>(games, HttpStatus.OK);
@@ -43,6 +50,8 @@ public class UserGameController {
 
     @RequestMapping(value = "", params = { "userId", "id" }, method = RequestMethod.GET)
     public ResponseEntity<GameDescription> getUserGame(@RequestParam("userId") String userId, @RequestParam("id") long id) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         GameDescription game = userService.getUserGame(userId, id);
 
         if (game == null) {
@@ -55,6 +64,8 @@ public class UserGameController {
 
     @RequestMapping(value = "/code", params = { "userId", "id" }, method = RequestMethod.GET)
     public ResponseEntity<Integer> getUserGameCode(@RequestParam("userId") String userId, @RequestParam("id") long id) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         int code = userService.getUserGameCode(userId, id);
 
         if (code == -1) {
@@ -67,12 +78,16 @@ public class UserGameController {
 
     @RequestMapping(value = "/count", params = { "userId" }, method = RequestMethod.GET)
     public ResponseEntity<Long> getNumberOfUserGames(@RequestParam("userId") String userId) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         long numberOfUserGames = userService.getNumberOfUserGames(userId);
         return new ResponseEntity<>(numberOfUserGames, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/count", params = { "userId", "kind", "league" }, method = RequestMethod.GET)
     public ResponseEntity<Long> getNumberOfUserGames(@RequestParam("userId") String userId, @RequestParam("kind") String kind, @RequestParam("league") String league) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         league = ControllerUtils.decodeUrlParameter(league);
         long numberOfUserGames = userService.getNumberOfUserGames(userId, kind, league);
         return new ResponseEntity<>(numberOfUserGames, HttpStatus.OK);
@@ -80,6 +95,8 @@ public class UserGameController {
 
     @PostMapping("")
     public ResponseEntity<GameDescription> createUserGame(@Valid @RequestBody GameDescription game) {
+        if (UserId.VBR_USER_ID.equals(game.getUserId())) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         boolean result = userService.createUserGame(game);
 
         if (result) {
@@ -93,6 +110,8 @@ public class UserGameController {
 
     @PutMapping("")
     public ResponseEntity<GameDescription> updateUserGame(@Valid @RequestBody GameDescription game) {
+        if (UserId.VBR_USER_ID.equals(game.getUserId())) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         boolean result = userService.updateUserGame(game);
 
         if (result) {
@@ -106,6 +125,8 @@ public class UserGameController {
 
     @RequestMapping(value = "", params = { "userId", "id" }, method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteUserGame(@RequestParam("userId") String userId, @RequestParam("id") long id) {
+        if (UserId.VBR_USER_ID.equals(userId)) { return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+
         boolean result = userService.deleteUserGame(userId, id);
 
         if (result) {
